requirements.txt
pdf2image
pytesseract
mistralai
python-dotenv
Pillow
opencv-python